feat: Add new options to main menu
This commits introduces new options to the main menu in the Main class. The user now is able to get the total price of all the booked seats and get ticket information based on row. It also changes the bookSeats method in the BookingSystem class, returning a Map containing the row and seat selected by user. This way, the main menu is able to create a Ticket instance based on these values and then this Ticket instance to the ticketsList (ArrayList).