feat: Prevent duplicates in ArrayList containing tickets
This commit prevents the ArrayList from containing tickets with the same row and same seat numbers. When booking seats (case 2 in the main menu), a Ticket instace is created. A new logic is added to check if the new instance of the Ticket class contains the same row and same seat number than the tickets instances already contained inside the ArrayList.