feat: Add customer attribute to Ticket class
This commits adds the customer attribute to the Ticket class that will contain a Customer class instance. This allows that every ticket gets related to a customer. When displaying a ticket information, the customer information will be shown as well. The main menu reflect this change by asking customer information to the user before the options are displayed on screen.